feat: weigh modes, PDF redesign, help sheet via Remote Config, market refactor
Tính năng mới
- 3 chế độ phiếu cân per-card (Card schema +5 field, DB migration 16→17):
  + Tạp chất kg/% (impurityIsPercent)
  + Bao bì đơn-vị / cân-mẫu n-bao (bagMethodIsSampling + bagSampleCount + bagSampleTotalWeight)
  + Quy cách nhập KG SMALL <100kg / LARGE ≥100kg (weightInputMode)
  + WeighOptionsSheet trên TopBar tab Cân Lúa để toggle
  + RiceCalculator.calcNetWeightWithModes + unit tests
- HelpBottomSheet trên TopBar tab Cân Lúa: 5 tip + nút "Hướng dẫn sử dụng"
  (mở YouTube/video) + "Xem trang web". Cả 2 URL đọc từ Firebase Remote Config
  (tutorial_video_url + website_url) — đổi không cần build lại app.

Redesign PDF biên lai
- Bỏ verification section, giữ QR để rescan dữ liệu phiếu
- Logic thanh toán mới: cọc + đã trả vs tổng → hiển thị tiền dư/còn nợ + trạng
  thái settled.
- Bảng cân động: 5 cột (kg+tare+impurity+net+amount) nếu phiếu có bao/tạp,
  rút gọn 3 cột nếu không có.
- Thêm row tổng dưới bảng cân.

Refactor MarketScreen
- TabRow + HorizontalPager 2 trang: Tin tức + Bảng giá lúa
- WeatherWidget inline vào NewsPage (cuộn cùng news), bỏ khỏi outer scope
- Weather chỉ refresh ở News tab, không trigger từ PricesPage
- FAB "Đăng giá mới" chỉ hiện ở tab Bảng giá lúa

Khác
- FirebaseRemoteConfigManager: thêm StateFlow cho tutorial_video_url, website_url
- Sync Card 5 field mới qua FirestoreCard + SyncManager
- Strings 5 locales (vi/en/km/lo/zh-rCN): help_*, weigh_options_*, pdf_*
- AppColors / Theme tweaks
- Repository + ViewModel updates cho weigh modes + sync

+++ app/src/main/java/com/GiaThinh/canlua/ui/component/cardlist/CardListSummaryCard.kt
@@ -73,7 +73,7 @@ fun CardListSummaryCard(
                     color = AppColors.TextSecondary
                 )
                 Text(
-                    text = stringResource(R.string.card_list_money_vnd, numberFormat.format(totalAmount)),
+                    text = stringResource(R.string.card_list_money_vnd, numberFormat.format(totalAmount.toLong())),
                     style = MaterialTheme.typography.headlineMedium,
                     fontWeight = FontWeight.Bold,
                     color = AppColors.GreenPrimary
